+++ warmup/likelihood/R_py_loglikelihood_comparison.py
@@ -1,8 +1,9 @@
 import pandas as pd
 import pmhn._trees._io as io
-from pmhn._trees._backend import OriginalTreeMHNBackend
+from pmhn._trees._backend import OriginalTreeMHNBackend, LoglikelihoodSingleTree
 import csv
 import numpy as np
+import time
 
 
 def csv_to_numpy(file_path):
@@ -48,20 +49,24 @@ trees_500 = io.parse_forest(df_500, naming=naming)
 # calculate loglikelihoods
 log_vec_py_AML = np.empty(len(trees_AML))
 log_vec_py_500 = np.empty(len(trees_500))
+start_time = time.time()
 backend = OriginalTreeMHNBackend()
-
 for idx, tree in trees_AML.items():
     print(f"Processing tree {idx} of {len(trees_AML)}")
-    log_value = backend.loglikelihood(tree, theta_AML, sampling_rate)
+    tree_log = LoglikelihoodSingleTree(tree)
+    log_value = backend.loglikelihood(tree_log, theta_AML, sampling_rate)
     log_vec_py_AML[idx - 1] = log_value
     print(f"log_value: {log_value}")
 
 for idx, tree in trees_500.items():
     print(f"Processing tree {idx} of {len(trees_500)}")
-    log_value = backend.loglikelihood(tree, theta_500, sampling_rate)
+    tree_log = LoglikelihoodSingleTree(tree)
+    log_value = backend.loglikelihood(tree_log, theta_500, sampling_rate)
     log_vec_py_500[idx - 1] = log_value
     print(f"log_value: {log_value}")
-
+end_time = time.time()
+elapsed_time = end_time - start_time
+print(f"Time elapsed: {elapsed_time} seconds")
 # write Python loglikelihoods to CSV
 np.savetxt("likelihood_py/log_vec_py_AML.csv", log_vec_py_AML, delimiter=",")
 np.savetxt("likelihood_py/log_vec_py_500.csv", log_vec_py_500, delimiter=",")
