+++ warmup/likelihood/R_py_loglikelihood_comparison_geno.py
@@ -1,8 +1,9 @@
 import pandas as pd
 import pmhn._trees._io as io
-from pmhn._trees._backend_new_v2 import OriginalTreeMHNBackend, LoglikelihoodSingleTree
+from pmhn._trees._backend_geno import OriginalTreeMHNBackend, LoglikelihoodSingleTree
 import csv
 import numpy as np
+import time
 
 
 def csv_to_numpy(file_path):
@@ -48,20 +49,33 @@ trees_500 = io.parse_forest(df_500, naming=naming)
 # calculate loglikelihoods
 log_vec_py_AML = np.empty(len(trees_AML))
 log_vec_py_500 = np.empty(len(trees_500))
+
+start_time = time.time()
 backend = OriginalTreeMHNBackend()
+theta_AML_size = len(theta_AML)
+all_mut_AML = set(i + 1 for i in range(theta_AML_size))
 for idx, tree in trees_AML.items():
     print(f"Processing tree {idx} of {len(trees_AML)}")
     tree_log = LoglikelihoodSingleTree(tree)
-    log_value = backend.loglikelihood(tree_log, theta_AML, sampling_rate)
+    log_value = backend.loglikelihood(
+        tree_log, theta_AML, sampling_rate, theta_AML_size, all_mut_AML
+    )
     log_vec_py_AML[idx - 1] = log_value
     print(f"log_value: {log_value}")
-
+theta_500_size = len(theta_500)
+all_mut_500 = set(i + 1 for i in range(theta_500_size))
 for idx, tree in trees_500.items():
     print(f"Processing tree {idx} of {len(trees_500)}")
     tree_log = LoglikelihoodSingleTree(tree)
-    log_value = backend.loglikelihood(tree_log, theta_500, sampling_rate)
+    log_value = backend.loglikelihood(
+        tree_log, theta_500, sampling_rate, theta_500_size, all_mut_500
+    )
     log_vec_py_500[idx - 1] = log_value
     print(f"log_value: {log_value}")
+
+end_time = time.time()
+elapsed_time = end_time - start_time
+print(f"Time elapsed: {elapsed_time} seconds")
 # write Python loglikelihoods to CSV
 np.savetxt("likelihood_py/log_vec_py_AML.csv", log_vec_py_AML, delimiter=",")
 np.savetxt("likelihood_py/log_vec_py_500.csv", log_vec_py_500, delimiter=",")
